+++ R/sd.R
@@ -1,0 +1,88 @@
+#' Standard Deviation and Variance of a Stressed Model
+#' 
+#' Provides the standard deviation and variance of stressed 
+#'     model components (random variables) under the scenario weights. 
+#' 
+#' @inheritParams mean_stressed
+#' @param xCol    Vector, or character vector,(names of) the columns
+#'                of the underlying data of the \code{object}
+#'                (\code{default = "all"}). 
+#' 
+#' @return \code{sd_stressed}: Return the standard deviation of the \code{xCol}
+#'     component of the stressed model with weights \code{wCol}.
+#'     The quantity can be evaluated at a vector. 
+#' 
+#' @details \code{sd_stressed}: The standard deviation of 
+#'      a chosen model component, subject to the calculated scenario weights.
+#'      
+#' @describeIn sd_stressed Sample standard deviation of model components
+#' 
+#' @examples      
+#' ## example with a stress on VaR
+#' set.seed(0)
+#' x <- as.data.frame(cbind(
+#'   "normal" = rnorm(1000), 
+#'   "gamma" = rgamma(1000, shape = 2)))
+#' res1 <- stress(type = "VaR", x = x, 
+#'   alpha = c(0.9, 0.95), q_ratio = 1.05)
+#' ## stressed standard deviation
+#' sd_stressed(res1, xCol = "all", wCol = "all", base = TRUE)
+#' 
+#' ## stressed variance
+#' var_stressed(res1, xCol = "all", wCol = "all", base = TRUE)
+#' 
+#' @author Kent Wu
+#' 
+#' @seealso See \code{\link{mean_stressed}} for means of stressed model components,
+#' and \code{\link{cor_stressed}} for correlations between stressed model components. 
+#' 
+#' @export
+
+sd_stressed <- function(object, xCol = "all", wCol = "all", base=FALSE){
+  if (!is.SWIM(object)) stop("Object not of class 'SWIM'")
+  
+  mean_w <- mean_stressed(object, xCol, wCol, base)
+  cname <- colnames(mean_w)
+  rname <- rownames(mean_w)
+  
+  if (is.character(xCol) && xCol == "all") xCol <- 1:ncol(get_data(object))
+  x_data <- as.matrix(get_data(object, xCol = xCol))
+  if (is.character(wCol) && wCol == "all") wCol <- 1:ncol(get_weights(object))
+  new_weights <- as.matrix(get_weights(object)[ ,wCol])
+  
+  if (anyNA(x_data)) warning("x contains NA")
+  
+  if (base == TRUE){
+    old_weights <- matrix(rep(1, length(x_data[,1])), ncol = 1)
+    new_weights <- cbind(old_weights, new_weights)
+  }
+  
+  n <- dim(x_data)[1] # number of observations
+  m <- dim(mean_w)[1] # number of weights
+  d <- dim(mean_w)[2] # number of random variables
+  
+  temp <- do.call(rbind, lapply(1:m, function(i) (x_data - rep(mean_w[i,], each=n))^2))
+  dim(new_weights) <- c(n, m, 1)
+  dim(temp) <- c(n, m, d)
+  sd <- do.call(rbind, lapply(1:m, function(i) sqrt(t(new_weights[,i,]) %*% temp[,i,] /(n-1) )))
+  colnames(sd) <- cname
+  rownames(sd) <- rname
+  
+  return(sd)
+}
+
+#' @describeIn sd_stressed Sample variance of model components
+#' 
+#' @return \code{var_stressed}: Return the variance of the \code{xCol}
+#'     component of the stressed model with weights \code{wCol}.
+#'     The quantity can be evaluated at a vector. 
+#'     
+#' @details \code{var_stressed}: The variance of 
+#'      a chosen stressed model component, subject to the calculated scenario weights.
+#'
+#' @export
+
+var_stressed <- function(object, xCol = "all", wCol = "all", base=FALSE){
+  var <- (sd_stressed(object, xCol, wCol, base))^2
+  return(var)
+}